Updated old tests and moved babel config to babel.config.js to work with jest
