--- .babelrc
@@ -1,8 +1,0 @@
-{
-	"presets": [
-		"@babel/preset-env"
-	],
-	"plugins": [
-		"import-graphql"
-	]
-}
