+++ babel.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+	presets: [['@babel/preset-env', { targets: { node: 'current' } }]],
+	plugins: ['import-graphql']
+};
