+++ src/connector/users/createUser.test.js
@@ -1,5 +1,5 @@
 import SQLite from 'sqlite3';
-import populate from '../sqlite/populate';
+import populate from '../../sqlite/populate';
 import PasswordHash from 'password-hash';
 
 import CreateUser from './createUser';
